Add files via upload
The table of Bayesian evidence and best fit model parameters with associated uncertainties for both the planetary engulfment model and the flat model. The best fit model parameters are calculated by taking the mean of the posterior samples, and the uncertainty is determined by taking the standard deviation of these samples. The parameter 'offset_flat' refers to $\Delta [M\H]_{offset}$ and 'offset_planet' refers to $\delta$.